😂 (#4)
Fix Azure SWA build after changing the app structure.

The new structure does not creat an index.html file in dist (which Azure
needs). But it creates an empty index.html in dist/app which Azure will
serve instead of calling the backend. Both issues can be fixed by rm and
touch but this is not a nice solution.